backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
websockets
motor
bcrypt
python-jose[cryptography]
email-validator
google-genai
